Fix potentially big bug
The distance observations are initialized to be infinity, so there is a
chance that no line intersects with the distance line, making it so that
the observation is infinity, and severely messing up with the gradient
calculations. The fix, as a result, is to set it at a big enough number
like 1000 instead.